librification: Pull stbt.py code into `_stbt/core.py`
This is in preparation for having a `DeviceUnderTest` object and removing
all this horrible global state.  The goal is to make stbt more amenable
to fitting into other testing frameworks such as py.test or behave.

Compatiblity is preserved by keeping `stbt/__init__.py`, but replacing
all the public functions with forwarding functions.  Future commits will
modify `core.py` to introduce the `DeviceUnderTest` class.

I've not moved the docstrings into core.py to avoid that duplication. This
may have to be revisited if we choose to make `DeviceUnderTest` a public
API.

This is a part of "librificaiton" work, better allowing stbt to be used as
a library, by removing side-effects.
